Add ability to restrict the scope to the namespace
When using a shared cluster I have no permissions to deploy a ClusterRole
and a ClusterRoleBinding. But I can deploy KubeView inside the
namespace I want to visualize.

Known issue: The dropdown menu to select the namespace to visualize is
empty because kubeview has no access to list namespaces. Just enter the
namespace you deployed kubeview in and press enter.